backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
opencv-python==4.9.0.80
numpy<2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
ultralytics==8.1.0
aiortc==1.9.0
av==12.0.0
aiohttp==3.9.1